Переделал сервис сетевых запросов. Убрал лишний и задваивающийся код. Откорректировал запрос ближашего города.
