--- Travel_Schedule/1_Models/StationListModel.swift
@@ -1,14 +1,0 @@
-//
-//  StationListModel.swift
-//  Travel_Schedule
-//
-//  Created by Kaider on 26.02.2025.
-//
-
-import Foundation
-
-let defaultStations = [
-    "Киевский вокзал",
-    "Курский вокзал",
-    "Ленинградский вокзал"
-]
